+++ scripts/gopackage.sh
@@ -2,13 +2,15 @@
 
 set -e
 
-echo "==> Packaging binaries..."
-
 source $(dirname $0)/version.sh
+
+VERSION_NO_V=$(echo $VERSION | sed "s/^[v|V]//")
+
+echo "==> Packaging binaries version ${VERSION_NO_V} ..."
 
 DIST=$(pwd)/dist/artifacts
 
-mkdir -p $DIST/${VERSION} $DIST/latest
+mkdir -p $DIST/${VERSION}
 
 for i in build/bin/*; do
     if [ ! -e $i ]; then
@@ -16,7 +18,7 @@ for i in build/bin/*; do
     fi
 
     BASE=build/archive
-    DIR=${BASE}/terraform-provider-rancher2-${VERSION}
+    DIR=${BASE}/${VERSION}
 
     rm -rf $BASE
     mkdir -p $BASE $DIR
@@ -26,34 +28,19 @@ for i in build/bin/*; do
         EXT=.exe
     fi
 
-    cp $i ${DIR}/terraform-provider-rancher2${EXT}
-
-    arch=$(echo $i | cut -f2 -d_)
-    mkdir -p $DIST/${VERSION}/binaries/$arch
-    mkdir -p $DIST/latest/binaries/$arch
-    cp $i $DIST/${VERSION}/binaries/$arch/terraform-provider-rancher2${EXT}
-    if [ -z "${EXT}" ]; then
-        gzip -c $i > $DIST/${VERSION}/binaries/$arch/terraform-provider-rancher2.gz
-        xz -c $i > $DIST/${VERSION}/binaries/$arch/terraform-provider-rancher2.xz
-    fi
-
-    rm -rf $DIST/latest/binaries/$arch
-    mkdir -p $DIST/latest/binaries
-    cp -rf $DIST/${VERSION}/binaries/$arch $DIST/latest/binaries
+    cp $i ${DIR}/terraform-provider-rancher2_${VERSION}${EXT}
 
     (
-        cd $BASE
-        NAME=$(basename $i | sed 's/_/-/g')
-        if [ -z "$EXT" ]; then
-            tar cvzf $DIST/${VERSION}/${NAME}-${VERSION}.tar.gz .
-            cp $DIST/${VERSION}/${NAME}-${VERSION}.tar.gz $DIST/latest/${NAME}.tar.gz
-
-            tar cvJf $DIST/${VERSION}/${NAME}-${VERSION}.tar.xz .
-            cp $DIST/${VERSION}/${NAME}-${VERSION}.tar.xz $DIST/latest/${NAME}.tar.xz
-        else
-            NAME=$(echo $NAME | sed 's/'${EXT}'//g')
-            zip -r $DIST/${VERSION}/${NAME}-${VERSION}.zip *
-            cp $DIST/${VERSION}/${NAME}-${VERSION}.zip $DIST/latest/${NAME}.zip
-        fi
+        cd $DIR
+        NAME=$(basename $i | cut -f1 -d_)
+        ARCH=$(basename $i | cut -f2,3 -d_ | cut -f1 -d.)
+        ARCHIVE=${NAME}_${VERSION_NO_V}_${ARCH}.zip
+        echo "Packaging dist/artifacts/${VERSION}/${ARCHIVE} ..."
+        zip -q $DIST/${VERSION}/${ARCHIVE} *
     )
 done
+
+(
+    cd $DIST/${VERSION}/
+    shasum -a 256 * > terraform-provider-rancher2_${VERSION_NO_V}_SHA256SUMS
+)